pyproject.toml
[project]
name = "immuvis-pigvae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.4",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "wandb",
    "rootutils",
    "pre-commit",
    "rich",
    "pytest",
    "matplotlib>=3.10.3",
    "scikit-learn>=1.7.0",
    "plotly>=6.2.0",
    "einops>=0.8.1",
    "ruff>=0.12.2",
    "mypy>=1.11.2",
    "ipykernel>=6.29.5",
    "onnx>=1.19.1",
    "onnxruntime>=1.23.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    "notebooks/**",
    "src/models/components/metrics/**",
]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # import sorting
    "B",   # bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # bandit (security)
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "PERF",
    "RUF",
    "TID",
    "ISC",
    "COM",
    "DTZ",
]
ignore = [
    "E203",  # black/ruff formatting compatibility
    "E501",  # handled by formatter; keep long lines when needed
    "E402",  # allow module imports not at top of file
    "N812",  # allow lowercase module imported as non-lowercase alias
    "COM812", # trailing comma conflicts with ruff formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S301", "S403", "S404"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
check_untyped_defs = false
no_implicit_optional = false
disallow_untyped_defs = false
exclude = [
    "^notebooks/",
    "^src/models/components/metrics/",
]
